+++ backend/dockerfile
@@ -1,0 +1,23 @@
+# Dockerfile for Python Backend
+FROM python:3.12-slim
+
+# Set working directory
+WORKDIR /app
+
+# Copy requirements file
+COPY backend/requirements.txt .
+
+# Install dependencies
+RUN pip install --no-cache-dir -r requirements.txt
+
+# Copy the entire backend directory
+COPY backend/ .
+
+# Make sure Python can find your modules
+ENV PYTHONPATH=/app
+
+# Expose the port the app runs on
+EXPOSE 5000
+
+# Command to run the application
+CMD ["uvicorn", "app.App:app", "--host", "0.0.0.0", "--port", "5000"]
