+++ backend/requirements.txt
@@ -8,22 +8,33 @@ blinker==1.9.0
 certifi==2025.6.15
 charset-normalizer==3.4.2
 click==8.2.1
+distro==1.9.0
 fastapi==0.115.13
 Flask==2.3.2
 flask-cors==6.0.1
 frozenlist==1.7.0
 h11==0.16.0
+httpcore==1.0.9
+httpx==0.28.1
 idna==3.10
+iniconfig==2.1.0
 itsdangerous==2.2.0
 Jinja2==3.1.6
+jiter==0.10.0
 MarkupSafe==3.0.2
 multidict==6.5.1
 numpy==2.3.1
-openai==0.27.8
-pandas==2.0.3
+openai==1.91.0
+packaging==25.0
+pandas==2.3.0
+pluggy==1.6.0
 propcache==0.3.2
 pydantic==2.11.7
 pydantic_core==2.33.2
+Pygments==2.19.2
+pytest==8.4.1
+pytest-asyncio==1.0.0
+pytest-mock==3.14.1
 python-dateutil==2.9.0.post0
 python-dotenv==1.0.0
 pytz==2025.2
