+++ docker-compose.yml
@@ -1,0 +1,22 @@
+version: '3.8'
+
+services:
+  backend:
+    build:
+      context: .  # Path to your project root
+      dockerfile: backend/Dockerfile
+    ports:
+      - "5000:5000"
+    volumes:
+    - ./backend/app/database.db:/app/app/database.db  # Mount the SQLite database file
+    environment:
+      - DATABASE_URL=sqlite:///app/database.db  # Adjust if necessary
+
+  frontend:
+    build:
+      context: .  # Path to your project root
+      dockerfile: frontend/Dockerfile
+    ports:
+      - "3000:3000"
+    depends_on:
+      - backend
