+++ frontend/dockerfile
@@ -1,0 +1,23 @@
+# Dockerfile for JavaScript Frontend
+FROM node:20
+
+# Set working directory
+WORKDIR /app
+
+# Copy package.json and package-lock.json
+COPY frontend/package*.json ./
+
+# Install dependencies
+RUN npm install
+
+# Copy the application code
+COPY frontend/ .
+
+# Build the application (if applicable)
+RUN npm run build
+
+# Expose the port the app runs on
+EXPOSE 3000
+
+# Command to run the application
+CMD ["npm", "start"]
